kernel: Link against libmstpm
The libmstpm.a is linked against the kernel only when the mstpm cargo feature
is enabled; currently it is enabled by default.

Tracking cargo features in the Makefile doesn't seem an easy/simple task, maybe
that's why we are not doing that (Issue #65). So, for now the libmstpm is
always built.

"make distclean" can be used to clean the libmstpm. Later that could be
improved by forcing the libmstpm to be built in the $OUT_DIR directory (place
in the target directory that cargo uses to build packages). With that the
libmstpm would also be cleaned when "make clean" is issued to delete the target
directory.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>