kernel/tdx/vmsr: Add vMSR framework
Add the Rust-based vMSR framework. This framework use trait MsrEmulation
which implements vMSR read/write and trait BoxedMsrEmulation which
implements the creation of vMSR objects. UnsupportedMsr is the default
case for vMSR emulation, erroring out on any accesses. This commit is
mainly to provide the skeleton for the upcoming commits.

Signed-off-by: Peter Fang <peter.fang@intel.com>