kernel/tdx/vmexit: Handle tdvmcall MMIO request
An enlightened TDP guest may use tdvmcall to access its MMIO, and this
MMIO request is handled by IoReq. According to the GHCI spec, MMIO read
is emulated by writing the read data back to R11 and MMIO write is
emulated by getting the data to be written from R15.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>