kernel/tdx/msr_bitmap: Add MSR bitmap support
Add support for MSR bitmap for the TDP guest. This framework is not yet
functional but provides the skeleton for the upcoming commits.

MsrBitmap allocates a 4KB-aligned MSR-bitmap page. MsrBitmapRef is the
structure intended for MSR bitmap control. It's meant to have a limited
lifetime so that td_set_msr_bitmap() calls can be done lazily in its
destructor. Only the updated (dirtied) bits are written via TDG.VP.WR
calls.

Note that according to the TDP spec:

  The default value of the MSR exit bitmaps shadow page is all-1, i.e.,
  L2->L1 exit on all MSR accesses.

Signed-off-by: Peter Fang <peter.fang@intel.com>