kernel/tdx/utils: Add helper function to enter TDP guest
td_vp_enter is the helper function to enter TDP guest. It takes a 256
bytes aligned GPA address as the TDP guest state and returns an eunm to
represents various outputs. If it is due to TDP guest vmexit, returns
detailed exit information. If it is due to svsm itself has pending
interrupts without entering TDP guest, returns without exit information
If it is due to some error, returns the detailed error code.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>