kernel/tdx/vmexit: Handle MSR read/write vmexit
The MSR read vmexit is handled by reading the value from vmsrs and
set to the corresponding guest general purpose registers.

The MSR write vmexit is handled by reading the value from the
corresponding guest general purpose registers and set to vmsrs to
emulate.

The RIP should be skipped for both if vmexit is handled successfully.

Co-developed-by: Jason Chen CJ <jason.cj.chen@intel.com>
Co-developed-by: Peter Fang <peter.fang@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>