kernel/cpu/idt: Init IDT 256 entries for handling interrupts
With local APIC enabled, interrupts will be happened in svsm, e.g. timer
interrupts/ipi interrupts. To handle these interrupts, IDT entries
should be extended to 256. The 0 ~ 31 vector is handled as exception and
the 32 ~ 256 is handled as interrupts by the generic interrupt handler
which only does an eoi. The dedicated interrupt handlers can be added on
top of this.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>