kernel/tdx/gctx: Add ia32_efer to GuestCpuContext
ia32_efer represents the IA32_EFER MSR of TDP guest, which is
implemented by using RegCache. As it is a fully emulated MSR which is
preserved across vmenter and vmexit, no need to provide a readhw
operation but just a writehw is enough. As RegCache will cache the dirty
value in memory so before vmentry, load ia32_efer to make sure the dirty
value is synced to VMCS.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>