kernel/tdx/vmsr: Add PassthruVmsr
PassthruVmsr can be used for any MSR that is directly passed through to
the guest. According to the TDP spec:

  By default, in cases where for L1 it would inject a #VE, the TDX
  module emulates an L2 to L1 VM exit.

If a vmexit is generated for a passthru MSR (marked as always-1 in the
MSR bitmap by the TDX module), a TDG.VP.VMCALL<INSTRUCTION.RDMSR/WRMSR>
is issued to the host VMM. If an error is returned by the host VMM, a
GP(0) will be injected into the guest.

Signed-off-by: Peter Fang <peter.fang@intel.com>