kernel/tdx/instr_emul: Add instruction decoding support
To support emulating MMIO/PIO instructions, InstrEmul is add as an
instruction decoder and emulator (emulation support will be added in the
following patch). The method decode_instr takes GuestCpuContext and
instruction length as the input, to fetch the raw instruction opcodes
from the current guest RIP and deliver the decoded information as the
output. Exception #PF or #UD will be injected to TDP guest if decoding
is failed.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>