kernel/tdx/vcpuid: Add initial vcpuid support
vcpuid is used to emulate the cpuid instruction for TDP guest. The
design is to make vcpuid as per VM as most of the cpuid leaves are
the same among vcpus and also not-changable. So the Vcpuid objects
contains a BTreeMap to maintain these common not-changable cpuid
entries. For the per vcpu related cpuid leaves, they will be emulated
when handles the cpuid instruction vmexit.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>