kernel/tdx/virq: Add Virq to support NMI injection
Add a new module Virq to support NMI injection. When needs to inject an
NMI, set a pending request INJ_NMI in VcpuReqFlags, which will be
handled in the vcpu run loop by Virq module before vmenter. If the NMI
cannot be injected then enable IRQ window so that NMI can be injected
once the irq injection is allowed.

Co-developed-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>