kernel/tdx/vmexit: Handle IO instruction vmexit
The vmexit caused by Port IO instruction is emulated through IoReq. For
the non-string IO instruction, the read result from a virtual port is
written back to vcpu RAX and data to be written to a virtual port is got
from vcpu RAX. This is done through PioEmul which implements the IoEmul
trait. For string IO instruction, it is similar with MMIO instruction
which needs the InstrEmul to decode and emulate.

As all the virtual port devices are emulated by host, reading from
a virtual port or writing to a virtual port is forwarded to the host
through tdvmcall.

As string IO instruction has repeat prefix, the RIP will be advanced
when the emulation for this repeat instruction is completed.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>