kernel/tdx/vcpu_comm: Add VlapicRegsRead in VcpuCommBlock
Introduce the VlapicRegsRead trait to facilitate the reading of vlapic
registers. By utilizing this trait, vlapic can provide the register
reading capability with writing not allowed. This trait can be employed
by VcpuCommBlock to read vlapic registers when enabling SMP. In this
scenario, a vcpu can read the VlapicRegs of a remote vcpu by using
remote vcpu's Arc<dyn VlapicRegsRead> in its VcpuCommBlock. This is
particularly useful for calculating the destination when waking up or
injecting IRQs into a remote vcpu.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>