kernel/tdx/vcpu: re-inject possible missed L2 interrupt
when SVSM try to vp enter L2, there is case that pending interrupt in L1
which will break the vp enter. For such case, SVSM does not need handle
such vmexit as it's dummy, but we still need take care of the interrupt
which may not be injected to L2 yet.

we need follow the handle_request sequence to inject possible missed
interrupt, so this patch takes use of virq.check_undelivered_event logic
to re-inject possible missed interrupt.

Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>