libmstpm: add `-mno-red-zone` in CFLAGS
From System V x86-64 abi, the 128-byte read zone area beyond the location
pointed to by %rsp is considered to be reserved and shall not be modified
by signal or interrupt handlers.

As the red area is enabled by default by gcc, the interrupt handler in the
current implementation may disrupt the data within the area.

To avoid unpredictable behavior caused by this, disable the red zone by
adding `-mno-red-zone` in CFLAGS for building openssl and ms-tpm-20-ref.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>